pyproject.toml
[project]
name = "news-research-analysis"
version = "0.1.0"
description = "A Streamlit chatbot for research and analysis of news articles using RAG and Gemini."
authors = [
    { name="Your Name", email="your.email@example.com" }
]
dependencies = [
    "streamlit",
    "python-dotenv",
    "langchain",
    "langchain-community",
    "langchain-huggingface",
    "langchain-google-genai",
    "sentence-transformers",
    "faiss-cpu",
    "unstructured",
    "lxml"
]
requires-python = ">=3.10"